refactor: split onboard preinstalled skills
